% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{add.equal}
\alias{add.equal}
\title{Count Elements Equal to a Value}
\usage{
add.equal(x, y)
}
\arguments{
\item{x}{A numeric vector. The vector of values to be compared.}

\item{y}{A numeric value. The value to compare the elements of 
\code{x} against.}
}
\value{
An integer representing the number of elements in \code{x} 
that are equal to \code{y}.
}
\description{
This function counts the number of elements in \code{x} that are 
equal to the specified
value \code{y}. It performs an element-wise comparison between 
\code{x} and \code{y} and returns
the total count of elements that satisfy the condition.
}
\examples{
add.equal(c(1, 2, 3, 4), 3)  # Returns 1 (one element is equal to 3)
add.equal(c(10, 5, 3, 5), 5) # Returns 2 (two elements are equal to 5)
add.equal(c(1, 2, 3), 4)     # Returns 0 (no elements are equal to 4)

}
