% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{biggerEqual}
\alias{biggerEqual}
\title{Compare Two Numbers for Greater or Equal}
\usage{
biggerEqual(x, y)
}
\arguments{
\item{x}{A numeric value. This is the first number to be compared.}

\item{y}{A numeric value. This is the second number to be compared.}
}
\value{
A logical value. Returns `TRUE` if \code{x} is greater than or
 equal to \code{y}, and `FALSE` otherwise.
If either \code{x} or \code{y} is not numeric, the function returns an 
error message.
}
\description{
This function takes two numeric inputs and returns `TRUE` if the first 
number is greater than
or equal to the second number, and `FALSE` otherwise. It checks if the 
input values are numeric.
}
\examples{
biggerEqual(5, 3) # returns TRUE
biggerEqual(2, 4) # returns FALSE
biggerEqual(3, 3) # returns TRUE
biggerEqual('a', 3) # returns an error message

}
