% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{compare}
\alias{compare}
\title{Compare Methods Based on Given Function Across Data}
\usage{
compare(
  measures,
  num.rows,
  num.columns,
  names.columns,
  names.rows,
  names.all,
  data,
  FUN
)
}
\arguments{
\item{measures}{A list or vector of measures for comparison (not used 
in the function body).}

\item{num.rows}{An integer specifying the number of rows in the dataframe.}

\item{num.columns}{An integer specifying the number of columns 
in the dataframe.}

\item{names.columns}{A character vector of column names, representing 
different methods.}

\item{names.rows}{A character vector of row names, representing 
different methods.}

\item{names.all}{A character vector of all method names to be compared.}

\item{data}{A dataframe containing the data to be compared.}

\item{FUN}{A function that takes two numeric vectors and returns 
a numeric vector 
with the comparison result.}
}
\value{
A dataframe where each column represents the comparison 
results between methods.
}
\description{
This function compares methods based on a provided comparison function. 
It creates a matrix
for each method, where each element represents the result of the 
comparison between methods
for each row of the data. It then aggregates these results into a final 
dataframe.
}
\examples{
df <- data.frame(Method1 = c(1, 2, 3), Method2 = c(2, 2, 1))
compare(measures = NULL, num.rows = 3, num.columns = 2, 
        names.columns = c("Method1", "Method2"), 
        names.rows = c("Method1", "Method2"), 
        names.all = c("Method1", "Method2"), 
        data = df, 
        FUN = function(x, y) x > y)

}
