% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{compare.rows}
\alias{compare.rows}
\title{Compare Values Across DataFrame Rows}
\usage{
compare.rows(num.rows, num.columns, names.columns, names.rows, data, FUN)
}
\arguments{
\item{num.rows}{An integer specifying the number of rows in the dataframe.}

\item{num.columns}{An integer specifying the number of columns in 
the dataframe.}

\item{names.columns}{A character vector of column names for the 
resulting matrix.}

\item{names.rows}{A character vector of row names for the resulting matrix.}

\item{data}{A dataframe containing the data to be compared.}

\item{FUN}{A function that takes two numeric vectors and returns 
a numeric vector 
with the comparison result.}
}
\value{
A dataframe with the same number of rows as the input dataframe 
and columns 
corresponding to the column names in `names.columns`. Each element 
represents the 
aggregated comparison result for each row across the columns of 
the dataframe.
}
\description{
This function compares values across rows of a dataframe using a 
specified function. 
It calculates a matrix of comparison results for each pair of columns, 
then aggregates
these results row-wise.
}
\examples{
df <- data.frame(A = c(1, 2, 3), B = c(3, 2, 1))
compare.rows(num.rows = 3, num.columns = 2, names.columns = c("A", "B"), 
             names.rows = c("Row1", "Row2", "Row3"), data = df, 
             FUN = function(x, y) x - y)

}
