% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{equal}
\alias{equal}
\title{Compare Two Numbers for Equality}
\usage{
equal(x, y)
}
\arguments{
\item{x}{A numeric value. This is the first number to be compared.}

\item{y}{A numeric value. This is the second number to be compared.}
}
\value{
A logical value. Returns `TRUE` if \code{x} is equal to 
\code{y}, and `FALSE` otherwise.
If either \code{x} or \code{y} is not numeric, the function returns 
an error message.
}
\description{
This function compares two numeric inputs and returns `TRUE` if the 
first number is equal to
the second number, and `FALSE` otherwise. It checks if the input 
values are numeric.
}
\examples{
equal(3, 3) # returns TRUE
equal(4, 5) # returns FALSE
equal(6, 6) # returns TRUE
equal('a', 3) # returns an error message

}
