% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{extract_measure_names}
\alias{extract_measure_names}
\title{Extract Measure Names from File Paths}
\usage{
extract_measure_names(file_paths)
}
\arguments{
\item{file_paths}{A character vector of full paths to the files.}
}
\value{
A character vector containing the names of the measures extracted 
from the file paths. Each name corresponds to the base name of a file 
without its extension.
}
\description{
This function extracts measure names from a vector of file paths by removing
the file extension from each file path. The measure names are assumed to be
the base names of the files without their extensions.
}
\examples{
# Example file paths
file_paths <- c(
  "C:/Users/Cissa/Documents/pairComparison/data/accuracy.csv",
  "C:/Users/Cissa/Documents/pairComparison/data/clp.csv"
)

# Extract measure names
measure_names <- extract_measure_names(file_paths)
print(measure_names)
# Output: "accuracy" "clp"

}
