% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{less}
\alias{less}
\title{Compare Two Numbers for Less Than}
\usage{
less(x, y)
}
\arguments{
\item{x}{A numeric value. This is the first number to be compared.}

\item{y}{A numeric value. This is the second number to be compared.}
}
\value{
A logical value. Returns `TRUE` if \code{x} is less than 
\code{y}, and `FALSE` otherwise.
If either \code{x} or \code{y} is not numeric, the function returns 
an error message.
}
\description{
This function compares two numeric inputs and returns `TRUE` if the 
first number is less than
the second number, and `FALSE` otherwise. It checks if the input values 
are numeric.
}
\examples{
less(3, 5) # returns TRUE
less(4, 4) # returns FALSE
less(6, 2) # returns FALSE
less('a', 3) # returns an error message

}
