% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{matrix.columns}
\alias{matrix.columns}
\title{Create a Square Matrix with Specified Size and Names}
\usage{
matrix.columns(names.columns, num.columns)
}
\arguments{
\item{names.columns}{A character vector of column names. This vector 
will also be used for row names.}

\item{num.columns}{An integer specifying the number of columns (and rows) 
in the matrix.}
}
\value{
A square matrix with dimensions \code{num.columns} 
by \code{num.columns}, with the specified
column and row names, and initialized with zeros.
}
\description{
This function constructs a square matrix with the specified number 
of columns and assigns
the provided column names as both row and column names. The matrix is 
initialized with zeros.
}
\examples{
colnames <- c("A", "B", "C")
mat <- matrix.columns(names.columns = colnames, num.columns = 3)
print(mat)

}
