% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{matrix.columns.rows}
\alias{matrix.columns.rows}
\title{Create a Matrix with Specified Dimensions and Names}
\usage{
matrix.columns.rows(names.columns, names.rows, num.rows, num.columns)
}
\arguments{
\item{names.columns}{A character vector of column names.}

\item{names.rows}{A character vector of row names.}

\item{num.rows}{An integer specifying the number of rows in the matrix.}

\item{num.columns}{An integer specifying the number of columns
in the matrix.}
}
\value{
A matrix with dimensions \code{num.rows} by 
\code{num.columns}, with the specified
column and row names, and initialized with zeros.
}
\description{
This function constructs a matrix with the specified number
 of rows and columns, and assigns
the provided column and row names. The matrix is initialized with zeros.
}
\examples{
colnames <- c("A", "B", "C")
rownames <- c("Row1", "Row2")
mat <- matrix.columns.rows(names.columns = colnames, 
names.rows = rownames, num.rows = 2, num.columns = 3)
print(mat)

}
