% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{matrix.rows.columns}
\alias{matrix.rows.columns}
\title{Create a Matrix with Specified Row Names and Column Names}
\usage{
matrix.rows.columns(names.rows, names.columns, num.rows, num.columns)
}
\arguments{
\item{names.rows}{A character vector of row names.}

\item{names.columns}{A character vector of column names.}

\item{num.rows}{An integer specifying the number of rows in the matrix.}

\item{num.columns}{An integer specifying the number of columns in the matrix.}
}
\value{
A matrix with dimensions \code{num.rows} by \code{num.columns}, with the specified
row names and column names, and initialized with zeros.
}
\description{
This function constructs a matrix with the specified number of rows and columns, and
assigns the provided row names and column names. The matrix is initialized with zeros.
}
\examples{
rownames <- c("Row1", "Row2", "Row3")
colnames <- c("Col1", "Col2")
mat <- matrix.rows.columns(names.rows = rownames, names.columns = colnames, 
                           num.rows = 3, num.columns = 2)
print(mat)

}
