% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{matrix.rows}
\alias{matrix.rows}
\title{Create a Square Matrix with Specified Size and Names}
\usage{
matrix.rows(names.rows, num.rows)
}
\arguments{
\item{names.rows}{A character vector of row names. This vector will
also be used for column names.}

\item{num.rows}{An integer specifying the number of rows (and columns) 
in the matrix.}
}
\value{
A square matrix with dimensions \code{num.rows} by \code{num.rows}, 
with the specified
row and column names, and initialized with zeros.
}
\description{
This function constructs a square matrix with the specified number of 
rows and assigns
the provided row names as both row and column names. The matrix is 
initialized with zeros.
}
\examples{
rownames <- c("X", "Y", "Z")
mat <- matrix.rows(names.rows = rownames, num.rows = 3)
print(mat)

}
