% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{subtraction}
\alias{subtraction}
\title{Subtract Two Numbers}
\usage{
subtraction(x, y)
}
\arguments{
\item{x}{A numeric value. This is the number from which
\code{y} will be subtracted.}

\item{y}{A numeric value. This is the number that will be 
subtracted from \code{x}.}
}
\value{
A numeric value. The result of \code{x} minus \code{y}. 
If either \code{x} or \code{y} 
is not numeric, the function throws an error.
}
\description{
This function subtracts the second numeric input from the first numeric 
input and returns the result.
It checks if the input values are numeric.
}
\examples{
subtraction(10, 5) # returns 5
subtraction(3, 7)  # returns -4
subtraction(4.5, 1.5) # returns 3
subtraction('a', 3) # throws an error

}
