% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{total.datasets.per.method}
\alias{total.datasets.per.method}
\title{Count Datasets Where One Method Performed Better Than Another}
\usage{
total.datasets.per.method(
  num.rows,
  num.columns,
  names.columns,
  names.rows,
  data,
  FUN
)
}
\arguments{
\item{num.rows}{An integer specifying the number of rows in the dataframe.}

\item{num.columns}{An integer specifying the number of columns in 
the dataframe.}

\item{names.columns}{A character vector of column names, representing 
different methods.}

\item{names.rows}{A character vector of row names, representing 
different methods.}

\item{data}{A dataframe containing the data to be compared.}

\item{FUN}{A function that takes two numeric vectors and returns a 
numeric vector 
with the comparison result.}
}
\value{
A list containing:
\item{matrix}{A matrix where each element represents the comparison 
result between methods.}
\item{mean.per.columns}{A dataframe with the mean comparison result 
per dataset (column-wise).}
\item{mean.per.rows}{A dataframe with the mean comparison result
 per method (row-wise).}
}
\description{
This function evaluates a comparison function across all pairs of 
columns in a dataframe.
It calculates how many times each method performed better than each 
other method
based on the comparison results. It also computes the mean performance 
per dataset
and per method.
}
\examples{
df <- data.frame(Method1 = c(1, 2, 3), Method2 = c(2, 2, 1))
total.datasets.per.method(num.rows = 3, num.columns = 2, 
                           names.columns = c("Method1", "Method2"), 
                           names.rows = c("Method1", "Method2"), 
                           data = df, 
                           FUN = function(x, y) sum(x > y))

}
