% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{total.performance.method}
\alias{total.performance.method}
\title{Calculate Performance of Methods Across Datasets}
\usage{
total.performance.method(
  folder.names.csv,
  Folder.CSVs,
  FolderMethods,
  my.methods,
  nomes.measures,
  res.mm
)
}
\arguments{
\item{folder.names.csv}{A character vector of CSV filenames containing 
the performance data.}

\item{Folder.CSVs}{A character string specifying the path to the folder 
containing the CSV files.}

\item{FolderMethods}{A character string specifying the path to the folder 
where results will be saved.}

\item{my.methods}{A character vector of method names used as column 
names in the data.}

\item{nomes.measures}{A character vector of measure names to process.}

\item{res.mm}{A list containing performance measures.}
}
\value{
A list containing the average performance of methods across datasets.
}
\description{
This function calculates the performance of methods across multiple 
datasets and measures. It compares
methods to determine how many other methods each method performed better 
than, for each dataset.
Results are saved as CSV files in the specified folder.
}
\examples{
# Assuming CSV files are available and directories are set correctly
result <- total.performance.method(
  folder.names.csv = c("measure1.csv", "measure2.csv"),
  Folder.CSVs = "data/csvs",
  FolderMethods = "data/methods",
  my.methods = c("Method1", "Method2"),
  nomes.measures = c("Measure1", "Measure2"),
  res.mm = list(measures = data.frame(names = c("Measure1", "Measure2"), 
  values = c(1, 0)))
)

}
