% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{bigger}
\alias{bigger}
\title{Compare Two Numbers for Greater Than}
\usage{
bigger(x, y)
}
\arguments{
\item{x}{A numeric value. This is the first number to be compared.}

\item{y}{A numeric value. This is the second number to be compared.}
}
\value{
A logical value. Returns `TRUE` if \code{x} is greater than 
\code{y}, and `FALSE` otherwise.
If either \code{x} or \code{y} is not numeric, the function returns 
an error message.
}
\description{
This function compares two numeric inputs and returns `TRUE` if the 
first number is greater than
the second number, and `FALSE` otherwise. It checks if the input values 
are numeric.
}
\examples{
bigger(5, 3) # returns TRUE
bigger(2, 4) # returns FALSE
bigger(3, 3) # returns FALSE
bigger('a', 3) # returns an error message

}
