% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{pc.plot.heatmap}
\alias{pc.plot.heatmap}
\title{Plot Heatmap from Comparison Results}
\usage{
pc.plot.heatmap(comparison_df, title = "Heatmap", desired_order)
}
\arguments{
\item{comparison_df}{A data frame containing the comparison results. The data frame should
be structured such that both rows and columns represent methods, with the values indicating 
the number of datasets in which one method was compared to another.}

\item{title}{A character string specifying the title of the heatmap.}

\item{desired_order}{A character vector specifying the desired order of methods for plotting.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function generates a heatmap from a dataframe where the values indicate 
the number of datasets in which methods were compared. It utilizes ggplot2 for visualization.
}
\examples{
# Example data frame (replace with your actual data)
comparison_df <- data.frame(
  Model_1 = c(14, 9, 5, 6),
  Model_2 = c(7, 14, 2, 4),
  Model_3 = c(9, 12, 14, 9),
  Model_4 = c(8, 10, 5, 14),
  row.names = c("Model_1", "Model_2", "Model_3", "Model_4")
)

# Plot heatmap
pc.plot.heatmap(comparison_df, title = "Comparison Heatmap", desired_order = c("Model_1", "Model_2", "Model_3", "Model_4"))

}
