% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{pc.plot.heatmap}
\alias{pc.plot.heatmap}
\title{Plot Heatmap from Comparison Results}
\usage{
pc.plot.heatmap(comparison_df, title = "Heatmap")
}
\arguments{
\item{comparison_df}{A data frame containing the comparison results. The data frame should
be in a format where the rows and columns are methods and the values represent the number of datasets
where one method was compared to another.}

\item{title}{A character string specifying the title of the heatmap.}
}
\value{
A ggplot object of the heatmap.
}
\description{
This function creates a heatmap from a dataframe where the values represent 
the number of datasets where methods were compared. It uses ggplot2 for visualization.
}
\examples{
# Example data frame (replace with your actual data)
comparison_df <- data.frame(
  Model_1 = c(14, 9, 5, 6),
  Model_2 = c(7, 14, 2, 4),
  Model_3 = c(9, 12, 14, 9),
  Model_4 = c(8, 10, 5, 14),
  row.names = c("Model_1", "Model_2", "Model_3", "Model_4")
)

# Plot heatmap
pc.plot.heatmap(comparison_df, title = "Comparison Heatmap")

}
