% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-comparison.R
\name{save.heatmap.as.pdf}
\alias{save.heatmap.as.pdf}
\title{Save a Heatmap Plot as a PDF File}
\usage{
save.heatmap.as.pdf(
  heatmap_plot,
  file_path,
  file_name,
  width = width,
  height = height
)
}
\arguments{
\item{heatmap_plot}{A ggplot2 object representing the heatmap plot to be saved.}

\item{file_path}{A character string specifying the directory where the PDF file will be saved.}

\item{file_name}{A character string specifying the name of the PDF file (without the extension).}

\item{width}{A numeric value specifying the width of the PDF file in inches (default is 10).}

\item{height}{A numeric value specifying the height of the PDF file in inches (default is 6).}
}
\value{
NULL. This function is called for its side effect of saving a PDF file.
}
\description{
This function saves a given heatmap plot as a PDF file in the specified directory.
}
\examples{
\dontrun{
# Example of saving a heatmap plot
heatmap_plot <- pc.plot.heatmap(comparison_df, title = "My Heatmap", order = desired_order)
save.heatmap.as.pdf(heatmap_plot, 
                    file_path = "C:/Users/Cissa/Documents/pairComparison/heatmaps", 
                    file_name = "heatmap_comparison")
}

}
